refactor: 라벨 생성, 수정, 삭제 시 권한확인

+++ backend/issue-tracker/src/main/java/com/issuetracker/core/label/domain/service/LabelCommandService.java
@@ -2,7 +2,10 @@ package com.issuetracker.core.label.domain.service;
 
 import com.issuetracker.core.label.domain.model.Label;
 import com.issuetracker.core.label.domain.port.LabelCommandPort;
+import com.issuetracker.core.label.domain.service.dto.CreateLabelInfo;
 import com.issuetracker.core.label.domain.service.dto.ModifyLabelInfo;
+import com.issuetracker.core.project.domain.service.dto.DeleteLabelInfo;
+import com.issuetracker.core.user.domain.model.User;
 import com.issuetracker.core.user.domain.service.UserQueryService;
 import jakarta.persistence.EntityExistsException;
 import lombok.RequiredArgsConstructor;
@@ -16,24 +19,34 @@ public class LabelCommandService {
 
     private final LabelCommandPort labelCommandPort;
     private final LabelQueryService labelQueryService;
+    private final UserQueryService userQueryService;
 
     // TODO: 2025-02-06  예외처리
-    public Label createLabel(String labelName) {
-        if (labelQueryService.isLabelPresentByName(labelName)) {
+    public Label createLabel(CreateLabelInfo info) {
+        User user = userQueryService.getUserById(info.userId());
+        user.validateAdminRole();
+
+        if (labelQueryService.isLabelPresentByName(info.name())) {
             throw new EntityExistsException();
         }
-        Label label = Label.create(labelName);
+        Label label = Label.create(info.name());
         return labelCommandPort.saveLabel(label);
     }
 
-    public Label modifyLabel(ModifyLabelInfo modifyLabelInfo) {
-        Label label = labelQueryService.getLabelById(modifyLabelInfo.id());
-        label.changeName(modifyLabelInfo.name());
+    public Label modifyLabel(ModifyLabelInfo info) {
+        User user = userQueryService.getUserById(info.userId());
+        user.validateAdminRole();
+
+        Label label = labelQueryService.getLabelById(info.labelId());
+        label.changeName(info.name());
         return labelCommandPort.saveLabel(label);
     }
 
-    public void deleteLabel(long id) {
-        Label label = labelQueryService.getLabelById(id);
+    public void deleteLabel(DeleteLabelInfo info) {
+        User user = userQueryService.getUserById(info.userId());
+        user.validateAdminRole();
+
+        Label label = labelQueryService.getLabelById(info.labelId());
         labelCommandPort.deleteLabel(label);
     }
 }
